// src/models/StarWarsCharacter.tsx
export interface StarWarsCharacter {
    name: string;
    height: number;
    mass: number;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: number;
    species: number;
    vehicles: number;
    starships: number;
    url: string;
}